Update document_tree
Still under heavy development.

Design decisions relate to incorporating document search functionality
and separating functions related to managing the trees from searching
documents.

In general the 'tree' has characteristics which make it more like a
document than a tree. There is a linear aspect to the tree, which allows
more freedom when writing some of the functions (I hope.)

The test suite is still a mess.

Functions being developed at the moment are:

	1. graft functions, which are going to graft a node onto a new
tree.
	2. trim functions, which are going to trim nodes from a tree.