Update document_search
There are 3 functions in document_search:

They are required for 2 doc_tree trimming functions:
	1. A trimming function that only trims nodes from trees
	2. A trimming function that trims sub-trees

	The first of these functions has been implemented in tag_search.
	tag_search returns all tags in the Doc_tree meeting the search
criteria.

	cont_search is simpler and returns all content nodes that
include the string str.

TODO: the second trimming function will require the search to break up
matching, so that only the roots of sub-trees will be returned, which
will then be needed to graft a new doc_tree.